[RL] Cleanup metrics: package split, output_suffix, TypedDict-style verification
- observability/metrics is now a package: metric_types.py + metric_logger.py
- inner MetricLogger.Config (with .build()) + MetricsConfig alias re-export
- LogprobVerification -> LogprobVerificationOutput dataclass; verify_logprob_identity
  returns local additive facts; trainer reducer assembles public bit_wise/* keys
- merge two reducers into reduce_forward_backward_metrics (one SUM pack + one MAX
  scalar; "loss" mesh; funcol.all_reduce kwargs; .detach() on metric numerators)
- rename suffix -> output_suffix on every reduction class + aggregator
- inline _population_std + _build_reward_metrics in grpo.py; delete grpo_metrics.py
- targeted fixes: stale annotation drop, B907 (!r) in checkpoint_path error,
  ASCII em-dashes
