+++ torchtitan/experiments/rl/observability/__init__.py
@@ -3,10 +3,3 @@
 #
 # This source code is licensed under the BSD-style license found in the
 # LICENSE file in the root directory of this source tree.
-
-"""Observability primitives for torchtitan RL.
-
-Currently exposes the typed metrics module. Import as::
-
-    from torchtitan.experiments.rl.observability import metrics as m
-"""
